Move LRUMap to browser
